Refactor: Create organized preferences folder structure
Architecture improvements:
- Create data/preferences/ folder for better organization
- Split preferences into focused classes:
  * CurrencyPreferences: currency selection & first launch
  * SummaryLayoutPreferences: layout type preferences
- Each preference class has its own SharedPreferences file
- Clean separation of concerns with logical grouping
- Updated all imports and references
- Removed old monolithic preference files

Result: Much cleaner, organized, and maintainable preference architecture
