+++ app/src/main/java/com/example/budget/data/AppContainer.kt
@@ -2,11 +2,13 @@ package com.example.budget.data
 
 import android.content.Context
 import com.example.budget.data.db.AppDatabase
+import com.example.budget.data.preferences.CurrencyPreferences
+import com.example.budget.data.preferences.SummaryLayoutPreferences
 
 interface AppContainer {
     val budgetRepository: BudgetRepository
     val currencyPreferences: CurrencyPreferences
-    val appPreferences: AppPreferences
+    val summaryLayoutPreferences: SummaryLayoutPreferences
 }
 
 class AppDataContainer(private val context: Context) : AppContainer {
@@ -22,7 +24,7 @@ class AppDataContainer(private val context: Context) : AppContainer {
         CurrencyPreferences(context)
     }
     
-    override val appPreferences: AppPreferences by lazy {
-        AppPreferences(context)
+    override val summaryLayoutPreferences: SummaryLayoutPreferences by lazy {
+        SummaryLayoutPreferences(context)
     }
 }
